feat: add location capacity management and authentication system
- Add capacity field to locations with default of 50
- Implement capacity update API endpoint and UI controls
- Add comprehensive authentication system with user roles and PIN login
- Implement encrypted credential storage using Security Crypto library
- Add Bluetooth printer support for label printing
- Enhance warehouse map with visual capacity indicators and core color visualization
- Add delivery day information to inventory screens
- Implement backend exception handling and connection status monitoring
- Add guest mode and auto-login functionality
- Reorganize home screen actions into categorized sections

+++ backend/src/main/kotlin/com/example/warehouse/repository/LocationRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface LocationRepository : JpaRepository<Location, Long> {
+interface LocationRepository : JpaRepository<Location, Int> {
     fun findByLabel(label: String): Location?
 }
